refactor: URL 생성 함수가 파라미터를 자유롭게 받아 동작할 수 있도록 개선
