+++ src/hooks/useProducts.ts
@@ -5,7 +5,7 @@ import { CategoryQueryString, PRODUCTS_SIZE, SortOptionQueryString } from "../co
 import useFetch from "./useFetch";
 import { Product } from "../types/products";
 import usePagination from "./usePagination";
-import createUrl from "../utils/createUrl";
+import { createFetchUrl } from "../utils";
 import { ENDPOINT } from "../constants/apis";
 
 interface UseProductResult {
@@ -30,15 +30,19 @@ export default function useProducts(): UseProductResult {
   const [category, setCategory] = useState<CategoryQueryString>("all");
   const [sortOption, setSortOption] = useState<SortOptionQueryString>("asc");
 
-  const url = createUrl({
-    endpoint: ENDPOINT.PRODUCT,
-    page: page === 0 ? page : page + PRODUCTS_SIZE.perRequest,
-    size: page === 0 ? PRODUCTS_SIZE.initial : PRODUCTS_SIZE.perRequest,
-    category,
+  const params = {
+    page: page === 0 ? page.toString() : (page + PRODUCTS_SIZE.perRequest).toString(),
+    size: page === 0 ? PRODUCTS_SIZE.initial.toString() : PRODUCTS_SIZE.perRequest.toString(),
+    category: category === "all" ? undefined : category,
     sortOption,
-  });
+  };
 
-  const { data, error, isLoading } = useFetch<ProductRequestResult>({ url, method: "GET" });
+  const url = createFetchUrl({ endpoint: ENDPOINT.PRODUCT, params });
+
+  const { data, error, isLoading } = useFetch<ProductRequestResult>({
+    url: url.href,
+    method: "GET",
+  });
 
   useEffect(() => {
     if (data) {
